refactor(merge): rename bDeletion → mergeDependents (B was spec jargon)
