+++ src/main/services/merge/mergeDependents.js
